Update dice_roller.py
added the veriable dice size, which asks the users for the number of sizes(number) that the dice they want to roll, and i also tweaked the veriable 'dice_rolls' to get an input from the user instead of being set to 2 every time. this is actually the number of dices that the program will roll.